move storagevalue to vec
