+++ src/agdb/collections.rs
@@ -1,5 +1,4 @@
 pub mod bit_set;
-pub mod db_vec;
 pub mod dictionary;
 pub mod dictionary_storage;
 pub mod indexed_map;
@@ -7,5 +6,6 @@ pub mod indexed_map_storage;
 pub mod map_storage;
 pub mod multi_map;
 pub mod multi_map_storage;
+pub mod vec;
 
 mod map;
